sysdeps_win32: declare [w]main without extern "C"
The C++ standard doesn't allow a linkage-specification for main,
https://eel.is/c++draft/basic.start.main#3.sentence-5:

"The main function shall not be declared with a linkage-specification
([dcl.link])."

Clang has started warning on incorrect declarations of main (-Wmain).

The MSDN docs show that wmain is also declared without 'extern "C"'.
After 'extern "C"' is removed, wmain is still emitted without mangling.
This special-casing is implemented in Clang for the IA-64 C++ ABI
in ItaniumMangleContextImpl::shouldMangleCXXName.

Bug: http://b/379133546
Test: m libadb_host
Test: manually verify that out/soong/.intermediates/packages/modules/adb/libadb_host/windows_x86_static/obj/packages/modules/adb/sysdeps_win32.o doesn't mangle wmain
Change-Id: Ib4963293314bce08d4196d439646941262c37f00